feat: Add header to to Protocol1 and single level hierarchy
The header currently has an optional tx_response channel which allows
the state machine to respond to "requests", such as the Protocol1::Get
message.

The single level hierarchy of states is accomplished by having the
state_process_msg_* return a boolean. If true is returned the message
was processed and false it was not. This provides a single level of
hierarchy in the state machine and allows processing of messaged common
to all states to processed in one location.