feat: Have fn_ptr states return StateResult
This is cleaner and has the benefit that transitioning can only be
done once per state function call. Previously the transition_to function
could be called any number of times which makes it ambiguous what
transitioning means and when it occurs. Now it always occurs upon
returning from the state function and only happens once.