feat: Added StateFns
Using StateFns and the Vector of StateFns allows dispatch_msg to be
general and it now contians no direct references to any of the State
functions.

To make this convenient to implemented I also declared the two constants
for the two indexes for the Vec<StateFns>.

Also moved fn do_transition into dispatch so it's not exposed.

There was a 40% decrease in performance because of the indirection,
so something to be aware of.

 dispatch_add_msg_to_state_add_fn_ptr_sm
                        time:   [4.4750 ns 4.4823 ns 4.4900 ns]
                        change: [+43.666% +43.983% +44.339%] (p = 0.00 < 0.05)
                        Performance has regressed.
 Found 4 outliers among 100 measurements (4.00%)
   1 (1.00%) high mild
   3 (3.00%) high severe

 dispatch_add_msg_to_state_any_fn_ptr_sm
                        time:   [4.3461 ns 4.3618 ns 4.3856 ns]
                        change: [+35.248% +35.766% +36.315%] (p = 0.00 < 0.05)
                        Performance has regressed.
 Found 16 outliers among 100 measurements (16.00%)
   9 (9.00%) high mild
   7 (7.00%) high severe

 dispatch_mul_msg_to_state_any_fn_ptr_sm
                        time:   [4.3410 ns 4.3587 ns 4.3841 ns]
                        change: [+37.742% +38.421% +39.090%] (p = 0.00 < 0.05)
                        Performance has regressed.
 Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild